move actions/ratingActions to actions/ratings
