+++ src/amo/actions/ratings.js
